+++ src/IKVM-pkg/buildTransitive/netstandard2.0/IKVM.targets
@@ -69,7 +69,7 @@
     <PropertyGroup>
         <GetIkvmReferenceItemsDependsOn>
             $(GetIkvmReferenceItemsDependsOn);
-            ResolveIkvmAssemblies;
+            ResolveIkvmReferenceAssemblies;
             GetIkvmReferenceItemsFromIkvmReferences;
             GetIkvmReferenceItemsMetadata;
         </GetIkvmReferenceItemsDependsOn>
@@ -91,7 +91,7 @@
     </Target>
 
     <!-- Builds the IkvmReferenceItem set into their output items within the cache. -->
-    <Target Name="_CompileIkvmReferences" DependsOnTargets="ResolveIkvmRuntimeAssembly;ResolveIkvmAssemblies;_GetIkvmReferenceItemsCompileToCachePath" Inputs="@(_IkvmReferenceItemCompileToCachePath)" Outputs="%(_IkvmReferenceItemCompileToCachePath.CachePath)" Condition=" '@(_IkvmReferenceItemCompileToCachePath.CachePath)' != '' ">
+    <Target Name="_CompileIkvmReferences" DependsOnTargets="ResolveIkvmRuntimeAssembly;ResolveIkvmReferenceAssemblies;_GetIkvmReferenceItemsCompileToCachePath" Inputs="@(_IkvmReferenceItemCompileToCachePath)" Outputs="%(_IkvmReferenceItemCompileToCachePath.CachePath)" Condition=" '@(_IkvmReferenceItemCompileToCachePath.CachePath)' != '' ">
 
         <!-- Take IkvmReferenceItem for the current batch on CachePath. -->
         <PropertyGroup>
@@ -143,7 +143,7 @@
         <CompileIkvmReferencesDependsOn>
             $(CompileIkvmReferenceAssemblies);
             GetIkvmReferenceItems;
-            ResolveIkvmAssemblies;
+            ResolveIkvmReferenceAssemblies;
             _GetIkvmReferenceItemsCompileToCachePath;
             _CompileIkvmReferences;
         </CompileIkvmReferencesDependsOn>
