+++ src/IKVM.MSBuild.Tests/IKVM.MSBuild.Tests.csproj
@@ -7,7 +7,7 @@
 
     <ItemGroup>
         <Compile Remove="Project\**\*" />
-        <None Include="Project\**\*" CopyToOutputDirectory="PreserveNewest" />
+        <None Include="Project\**\*" CopyToOutputDirectory="Always" />
     </ItemGroup>
 
     <ItemGroup>
