+++ src/IKVM.NET.Sdk.Tests/IKVM.NET.Sdk.Tests.csproj
@@ -7,7 +7,7 @@
 
     <ItemGroup>
         <Compile Remove="Project\**\*" />
-        <None Include="Project\**\*" CopyToOutputDirectory="PreserveNewest" />
+        <None Include="Project\**\*" CopyToOutputDirectory="Always" />
     </ItemGroup>
 
     <ItemGroup>
