+++ src/IKVM.NET.Sdk.Tests/Project/Lib/ProjectLib.msbuildproj
@@ -6,7 +6,6 @@
 
     <PropertyGroup>
         <TargetFrameworks>net461;netcoreapp3.1</TargetFrameworks>
-        <RuntimeIdentifiers>win7-x64;linux-x64</RuntimeIdentifiers>
     </PropertyGroup>
     
 </Project>
