+++ src/IKVM.NET.Sdk/targets/IKVM.Java.Core.Commands.targets
@@ -55,30 +55,30 @@
             <_IkvmStubArgs Include="-bootstrap" Condition=" '$(Bootstrap)' == 'true' " />
             <_IkvmStubArgs Include="@(ReferencePathWithRefAssemblies->'-r:&quot;%(FullPath)&quot;')"  Condition=" '%(Filename)' != '%(ReferenceExport.Filename)' "/>
             <_IkvmStubArgs Include="-nostdlib" Condition=" '$(NoCompilerStandardLib)' == 'true' " />
-            <_IkvmStubArgs Include="-out:&quot;$(ReferenceStubOutputPath)tmp\%(ReferenceExport.Filename)%(ReferenceExport.Extension)&quot;" />
+            <_IkvmStubArgs Include="-out:&quot;$(ReferenceExportOutputPath)tmp\%(ReferenceExport.Filename)%(ReferenceExport.Extension)&quot;" />
             <_IkvmStubArgs Include="&quot;%(ReferenceExport.AssemblyFile)&quot;" />
         </ItemGroup>
 
-        <MakeDir Directories="$(ReferenceStubOutputPath);$(ReferenceStubOutputPath)tmp" />
+        <MakeDir Directories="$(ReferenceExportOutputPath);$(ReferenceExportOutputPath)tmp" />
         <Message Text="$(IkvmExporterExec) @(_IkvmStubArgs, ' ')" Importance="low" />
         <Exec Command="$(IkvmExporterExec) @(_IkvmStubArgs, ' ')" />
-        <Move SourceFiles="$(ReferenceStubOutputPath)tmp\%(ReferenceExport.Filename)%(ReferenceExport.Extension)" DestinationFiles="%(ReferenceExport.Identity)" />
-        <Delete Files="$(ReferenceStubOutputPath)tmp\%(ReferenceExport.Filename)%(ReferenceExport.Extension)" />
+        <Move SourceFiles="$(ReferenceExportOutputPath)tmp\%(ReferenceExport.Filename)%(ReferenceExport.Extension)" DestinationFiles="%(ReferenceExport.Identity)" />
+        <Delete Files="$(ReferenceExportOutputPath)tmp\%(ReferenceExport.Filename)%(ReferenceExport.Extension)" />
 
         <ItemGroup>
-            <FileWrites Include="$(ReferenceStubOutputPath)tmp\%(ReferenceExport.Filename)%(ReferenceExport.Extension)" />
+            <FileWrites Include="$(ReferenceExportOutputPath)tmp\%(ReferenceExport.Filename)%(ReferenceExport.Extension)" />
         </ItemGroup>
     </Target>
 
-    <Target Name="ResolveJavaCompilerPath">
+    <Target Name="ResolveJavaCompiler">
         <PropertyGroup>
-            <JavaCompilerPath>javac</JavaCompilerPath>
-            <JavaCompilerArgs>-J-Xmx1536M</JavaCompilerArgs>
-            <JavaCompilerExec>$(JavaCompilerPath) $(JavaCompilerArgs)</JavaCompilerExec>
+            <JavaCompilerPath Condition=" '$(JavaCompilerPath)' == '' ">javac</JavaCompilerPath>
+            <JavaCompilerArgs Condition=" '$(JavaCompilerArgs)' == '' ">-J-Xmx1536M</JavaCompilerArgs>
+            <JavaCompilerExec Condition=" '$(JavaCompilerExec)' == '' ">$(JavaCompilerPath) $(JavaCompilerArgs)</JavaCompilerExec>
         </PropertyGroup>
     </Target>
 
-    <Target Name="_CompileJava" DependsOnTargets="ResolveJavaCompilerPath;BuildReferenceStubs" Inputs="@(Compile);@(Classpath)" Outputs="$(_CompileJavaStampFile)">
+    <Target Name="_CompileJava" DependsOnTargets="ResolveJavaCompiler;BuildReferenceExports" Inputs="@(Compile);@(Classpath)" Outputs="$(_CompileJavaStampFile)">
         <ItemGroup>
             <_JavaCompilerArgs Include="-g" Condition=" '$(DebugSymbols)' == 'true' Or '$(DebugType)' != 'none' " />
             <_JavaCompilerArgs Include="-nowarn" />
